fix: resolve Unity 6 compilation errors and enable headless testing
## Compilation Fixes
- Fixed Cinemachine namespace (Unity.Cinemachine)
- Fixed Input namespace collisions (UnityEngine.Input.GetKeyDown)
- Added missing using statements (IronFrontier.Core, IronFrontier.Systems)
- Added Wind module to package manifest
- Fixed URP VolumeProfile references

## Assembly Structure
- Created IronFrontier.Runtime.asmdef for main game code
- Removed redundant sub-assembly definitions
- Moved tests to Assets/Tests/ for proper Unity discovery
- Updated test assemblies to reference main game assembly

## Interface Fixes (via parallel agents)
- Added Gold, AddGold, RemoveGold to InventoryManager
- Added CanInteract, LookAt to PlayerController
- Added ToState to PlayerStateChangedEventArgs
- Added Disabled to PlayerState enum
- Fixed SceneController.sceneTransition access level
- Fixed float/int casts in CombatSceneLoader
- Fixed yield in try-catch in GameBootstrap

## Test Results
- 165 tests discovered
- 153 passed (92.7%)
- 12 failed (quest event edge cases)

All changes support fully headless Unity development via batch mode.

+++ Assets/Scripts/Combat/CombatManager.cs
@@ -757,7 +757,7 @@ namespace IronFrontier.Combat
             var enemies = GetLivingEnemies();
 
             float avgEnemySpeed = enemies.Count > 0
-                ? enemies.Average(e => e.Stats.Speed)
+                ? (float)enemies.Average(e => e.Stats.Speed)
                 : actorStats.Speed;
 
             float speedDiff = actorStats.Speed - avgEnemySpeed;
